Add some escaping for the name and search fields.
Certain content could break TagBar parsing, such as having a backslash
in the title.  It's okay to have an unescaped backslash in the name
field, but not the search field.

So, we process the names to remove other kinds of whitespace and simply
replace them with a single space.  Then we escape the search string so
that the location can actually be found by Vim and others.