Fix #7: wrong tag section layout
This employs a better method for maintain the section paths.  Instead of
trying to treat previousSections as an array where each entry
represented the correct level heading, we instead compact it and
recognize that there may be gaps in levels.  Now we pop off any level
greater than or equal to the current level, and use the remaining data
to formulate the path.