Cope with encoding issues in Python 3.
We don't really know the encoding of the input file, so we must do our
best to cope with whatever is present.  As a result, we'll treat the
file as latin1 under Python 3 to avoid losing any byte content while
processing the data, since all byte values are valid for latin1.