Fix #11: Cannot use empty key for Dictionary
While I could not produce the exact error specified in the issue, the
bottom line is that the pipe symbol (`|`) is being used to do scope
resolution (the option is called "sro").  My understanding was that this
only applied to latter half of the tagline, but it appears TagBar is
splitting on the sro in the first part containing the tagname causing
"Logical Operators ( || )" to be seen as a multipart reference to a tag.

To prevent regressing other users, we'll keep the default sro character
but allow a different one to be specified.